backend/requirements.txt
fastapi
uvicorn
transformers
pytesseract
Pillow
PyMuPDF
torch
torchvision
torchaudio
tensorflow-intel==2.16.2
transformers==4.39.3
keras
tensorboard<2.17
python-multipart